+++ packages/ai-constructs/src/conversation/runtime/conversation_turn_response_sender.test.ts
@@ -5,12 +5,17 @@ import {
   MutationResponseInput,
   MutationStreamingResponseInput,
 } from './conversation_turn_response_sender';
-import { ConversationTurnEvent, StreamingResponseChunk } from './types';
+import {
+  ConversationTurnError,
+  ConversationTurnEvent,
+  StreamingResponseChunk,
+} from './types';
 import { ContentBlock } from '@aws-sdk/client-bedrock-runtime';
 import {
   GraphqlRequest,
   GraphqlRequestExecutor,
 } from './graphql_request_executor';
+import { UserAgentProvider } from './user_agent_provider';
 
 void describe('Conversation turn response sender', () => {
   const event: ConversationTurnEvent = {
@@ -33,7 +38,19 @@ void describe('Conversation turn response sender', () => {
   };
 
   void it('sends response back to appsync', async () => {
-    const graphqlRequestExecutor = new GraphqlRequestExecutor('', '', '');
+    const userAgentProvider = new UserAgentProvider(
+      {} as unknown as ConversationTurnEvent
+    );
+    const userAgentProviderMock = mock.method(
+      userAgentProvider,
+      'getUserAgent',
+      () => 'testUserAgent'
+    );
+    const graphqlRequestExecutor = new GraphqlRequestExecutor(
+      '',
+      '',
+      userAgentProvider
+    );
     const executeGraphqlMock = mock.method(
       graphqlRequestExecutor,
       'executeGraphql',
@@ -43,6 +60,7 @@ void describe('Conversation turn response sender', () => {
     );
     const sender = new ConversationTurnResponseSender(
       event,
+      userAgentProvider,
       graphqlRequestExecutor
     );
     const response: Array<ContentBlock> = [
@@ -53,7 +71,14 @@ void describe('Conversation turn response sender', () => {
     ];
     await sender.sendResponse(response);
 
+    assert.strictEqual(userAgentProviderMock.mock.calls.length, 1);
+    assert.deepStrictEqual(userAgentProviderMock.mock.calls[0].arguments[0], {
+      'turn-response-type': 'single',
+    });
     assert.strictEqual(executeGraphqlMock.mock.calls.length, 1);
+    assert.deepStrictEqual(executeGraphqlMock.mock.calls[0].arguments[1], {
+      userAgent: 'testUserAgent',
+    });
     const request = executeGraphqlMock.mock.calls[0]
       .arguments[0] as GraphqlRequest<MutationResponseInput>;
     assert.deepStrictEqual(request, {
@@ -81,7 +106,15 @@ void describe('Conversation turn response sender', () => {
   });
 
   void it('serializes tool use input to JSON', async () => {
-    const graphqlRequestExecutor = new GraphqlRequestExecutor('', '', '');
+    const userAgentProvider = new UserAgentProvider(
+      {} as unknown as ConversationTurnEvent
+    );
+    mock.method(userAgentProvider, 'getUserAgent', () => '');
+    const graphqlRequestExecutor = new GraphqlRequestExecutor(
+      '',
+      '',
+      userAgentProvider
+    );
     const executeGraphqlMock = mock.method(
       graphqlRequestExecutor,
       'executeGraphql',
@@ -91,6 +124,7 @@ void describe('Conversation turn response sender', () => {
     );
     const sender = new ConversationTurnResponseSender(
       event,
+      userAgentProvider,
       graphqlRequestExecutor
     );
     const toolUseBlock: ContentBlock.ToolUseMember = {
@@ -136,7 +170,19 @@ void describe('Conversation turn response sender', () => {
   });
 
   void it('sends streaming response chunk back to appsync', async () => {
-    const graphqlRequestExecutor = new GraphqlRequestExecutor('', '', '');
+    const userAgentProvider = new UserAgentProvider(
+      {} as unknown as ConversationTurnEvent
+    );
+    const userAgentProviderMock = mock.method(
+      userAgentProvider,
+      'getUserAgent',
+      () => 'testUserAgent'
+    );
+    const graphqlRequestExecutor = new GraphqlRequestExecutor(
+      '',
+      '',
+      userAgentProvider
+    );
     const executeGraphqlMock = mock.method(
       graphqlRequestExecutor,
       'executeGraphql',
@@ -146,6 +192,7 @@ void describe('Conversation turn response sender', () => {
     );
     const sender = new ConversationTurnResponseSender(
       event,
+      userAgentProvider,
       graphqlRequestExecutor
     );
     const chunk: StreamingResponseChunk = {
@@ -158,7 +205,14 @@ void describe('Conversation turn response sender', () => {
     };
     await sender.sendResponseChunk(chunk);
 
+    assert.strictEqual(userAgentProviderMock.mock.calls.length, 1);
+    assert.deepStrictEqual(userAgentProviderMock.mock.calls[0].arguments[0], {
+      'turn-response-type': 'streaming',
+    });
     assert.strictEqual(executeGraphqlMock.mock.calls.length, 1);
+    assert.deepStrictEqual(executeGraphqlMock.mock.calls[0].arguments[1], {
+      userAgent: 'testUserAgent',
+    });
     const request = executeGraphqlMock.mock.calls[0]
       .arguments[0] as GraphqlRequest<MutationStreamingResponseInput>;
     assert.deepStrictEqual(request, {
@@ -177,7 +231,15 @@ void describe('Conversation turn response sender', () => {
   });
 
   void it('serializes tool use input to JSON when streaming', async () => {
-    const graphqlRequestExecutor = new GraphqlRequestExecutor('', '', '');
+    const userAgentProvider = new UserAgentProvider(
+      {} as unknown as ConversationTurnEvent
+    );
+    mock.method(userAgentProvider, 'getUserAgent', () => '');
+    const graphqlRequestExecutor = new GraphqlRequestExecutor(
+      '',
+      '',
+      userAgentProvider
+    );
     const executeGraphqlMock = mock.method(
       graphqlRequestExecutor,
       'executeGraphql',
@@ -187,6 +249,7 @@ void describe('Conversation turn response sender', () => {
     );
     const sender = new ConversationTurnResponseSender(
       event,
+      userAgentProvider,
       graphqlRequestExecutor
     );
     const toolUseBlock: ContentBlock.ToolUseMember = {
@@ -232,6 +295,83 @@ void describe('Conversation turn response sender', () => {
               },
             },
           ],
+        },
+      },
+    });
+  });
+
+  void it('sends errors response back to appsync', async () => {
+    const userAgentProvider = new UserAgentProvider(
+      {} as unknown as ConversationTurnEvent
+    );
+    const userAgentProviderMock = mock.method(
+      userAgentProvider,
+      'getUserAgent',
+      () => 'testUserAgent'
+    );
+    const graphqlRequestExecutor = new GraphqlRequestExecutor(
+      '',
+      '',
+      userAgentProvider
+    );
+    const executeGraphqlMock = mock.method(
+      graphqlRequestExecutor,
+      'executeGraphql',
+      () =>
+        // Mock successful Appsync response
+        Promise.resolve()
+    );
+    const sender = new ConversationTurnResponseSender(
+      event,
+      userAgentProvider,
+      graphqlRequestExecutor
+    );
+    const errors: Array<ConversationTurnError> = [
+      {
+        errorType: 'errorType1',
+        message: 'errorMessage1',
+      },
+      {
+        errorType: 'errorType2',
+        message: 'errorMessage2',
+      },
+    ];
+    await sender.sendErrors(errors);
+
+    assert.strictEqual(userAgentProviderMock.mock.calls.length, 1);
+    assert.deepStrictEqual(userAgentProviderMock.mock.calls[0].arguments[0], {
+      'turn-response-type': 'error',
+    });
+    assert.strictEqual(executeGraphqlMock.mock.calls.length, 1);
+    assert.deepStrictEqual(executeGraphqlMock.mock.calls[0].arguments[1], {
+      userAgent: 'testUserAgent',
+    });
+    assert.strictEqual(executeGraphqlMock.mock.calls.length, 1);
+    const request = executeGraphqlMock.mock.calls[0]
+      .arguments[0] as GraphqlRequest<MutationResponseInput>;
+    assert.deepStrictEqual(request, {
+      query:
+        '\n' +
+        '        mutation PublishModelResponse($input: testResponseMutationInputTypeName!) {\n' +
+        '            testResponseMutationName(input: $input) {\n' +
+        '                testSelectionSet\n' +
+        '            }\n' +
+        '        }\n' +
+        '    ',
+      variables: {
+        input: {
+          conversationId: event.conversationId,
+          errors: [
+            {
+              errorType: 'errorType1',
+              message: 'errorMessage1',
+            },
+            {
+              errorType: 'errorType2',
+              message: 'errorMessage2',
+            },
+          ],
+          associatedUserMessageId: event.currentMessageId,
         },
       },
     });