+++ packages/auth-construct/src/index.ts
@@ -1,6 +1,5 @@
 export {
   AuthProps,
-  ReferenceAuthProps,
   EmailLogin,
   EmailLoginSettings,
   GoogleProviderProps,
@@ -27,7 +26,7 @@ export {
   CustomAttributeBoolean,
   CustomAttributeDateTime,
   CustomAttributeBase,
+  CustomEmailSender,
 } from './types.js';
 export { AmplifyAuth } from './construct.js';
-export { AmplifyReferenceAuth } from './reference_construct.js';
 export { triggerEvents } from './trigger_events.js';