+++ packages/backend-auth/src/index.ts
@@ -3,5 +3,6 @@ export {
   BackendReferenceAuth,
   AmplifyReferenceAuthProps,
   referenceAuth,
+  ReferenceAuthProps,
 } from './reference_factory.js';
 export * from './types.js';