+++ packages/backend-auth/src/reference_construct.test.ts
@@ -3,15 +3,15 @@ import assert from 'assert';
 import {
   AmplifyReferenceAuth,
   OUTPUT_PROPERTIES_PROVIDED_BY_AUTH_CUSTOM_RESOURCE,
+  authOutputKey,
 } from './reference_construct.js';
-import { ReferenceAuthProps } from './types.js';
 import {
   BackendOutputEntry,
   BackendOutputStorageStrategy,
 } from '@aws-amplify/plugin-types';
-import { authOutputKey } from '@aws-amplify/backend-output-schemas';
 import { Template } from 'aws-cdk-lib/assertions';
 import { App, Stack } from 'aws-cdk-lib';
+import { ReferenceAuthProps } from './reference_factory.js';
 const refAuthProps: ReferenceAuthProps = {
   authRoleArn: 'arn:aws:iam::000000000000:role/amplify-sample-auth-role-name',
   unauthRoleArn: